inclusive-backend/requirements.txt
allauth
django-rest-framework
django-rest-framework-simplejwt
dj-rest-auth
django-cors-headers
django-allauth
django-allauth-ui
django-allauth-email
django-allauth-socialaccount
django-allauth-socialaccount-providers
django-allauth-socialaccount-providers-google
django-allauth-socialaccount-providers-facebook
django-allauth-socialaccount-providers-twitter
django-allauth-socialaccount-providers-linkedin
django-allauth-socialaccount-providers-github
django-allauth-socialaccount-providers-apple
django-allauth-socialaccount-providers-microsoft
django-allauth-socialaccount-providers-yahoo
django-allauth-socialaccount-providers-amazon
django-allauth-socialaccount-providers-discord
django-allauth-socialaccount-providers-spotify
django-allauth-socialaccount-providers-reddit
django-allauth-socialaccount-providers-twitch
django-allauth-socialaccount-providers-apple
django-allauth-socialaccount-providers-microsoft
django-allauth-socialaccount-providers-yahoo
django-allauth-socialaccount-providers-amazon
django-allauth-socialaccount-providers-discord
django-allauth-socialaccount-providers-spotify
django-allauth-socialaccount-providers-reddit
django-allauth-socialaccount-providers-twitch
django-allauth-socialaccount-providers-apple
django-allauth-socialaccount-providers-microsoft
django-allauth-socialaccount-providers-yahoo
django-allauth-socialaccount-providers-amazon
django-allauth-socialaccount-providers-discord
django-allauth-socialaccount-providers-spotify
django-allauth-socialaccount-providers-reddit
django-allauth-socialaccount-providers-twitch

